refactor(FNP-45): update import paths and configurations for consistency and maintainability
- Updated import paths to use absolute paths in SignUpPage, SignInPage, ErrorBoundaryPage, and auth-api for consistency
- Renamed `error-doundary.tsx` to `error-boundary.tsx` for correct spelling
- Added alias for `src` directory in Vite configuration for cleaner import paths
- Modified tsconfig.json to include `baseUrl` and `paths` for using `src` alias
- Rearranged and updated compiler options in tsconfig.json for better configuration management
