+++ packages/client/src/pages/ErrorBoundaryPage/ErrorBoundaryPage.tsx
@@ -1,5 +1,5 @@
 import { FC, useEffect } from 'react';
-import ErrorBoundary from '../../services/error/error-doundary';
+import ErrorBoundary from 'src/services/error/error-boundary';
 
 const NormalComponent: FC = () => {
   return <div>normal</div>;
