+++ packages/client/src/pages/SignUpPage/SignUpPage.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
-import { TSignUpRequest, TUser } from '../../shared/types/user';
-import AuthAPI from '../../services/api/auth-api';
+import { TSignUpRequest, TUser } from 'src/shared/types/user';
+import AuthAPI from 'src/services/api/auth-api';
 import { Link } from 'react-router-dom';
-import Helpers from '../../utils/helpers';
+import Helpers from 'src/utils/helpers';
 
 const authAPI = new AuthAPI();
