+++ packages/client/tsconfig.json
@@ -14,7 +14,11 @@
     "forceConsistentCasingInFileNames": true,
     "useDefineForClassFields": true,
     "isolatedModules": true,
-    "noEmit": true
+    "noEmit": true,
+    "baseUrl": ".",
+    "paths": {
+      "src": ["./src/*"]
+    }
   },
   "include": ["src"],
   "references": [{ "path": "./tsconfig.node.json" }]
